+++ test/helpers/application_helper_test.rb
@@ -28,22 +28,39 @@ class ApplicationHelperTest < ActionView::TestCase
                  title_tag_contents(action_name))
   end
 
-  def test_make_table
-    expect = "<table><tr><td>1</td><td>2</td></tr>" \
-             "<tr><td>3</td><td>4</td></tr></table>"
-    table = make_table([[1, 2], [3, 4]])
-    assert_equal(expect, table)
+  def test_add_args_to_url_two_args
+    assert_equal("/abcdef?foo=bar&this=that",
+                 add_args_to_url("/abcdef", foo: "bar", this: "that"))
   end
 
-  def test_make_table_with_colspan
-    expect = '<table><tr colspan="2"><td>5</td><td>6</td></tr></table>'
-    table = make_table([[5, 6]], {}, { colspan: 2 })
-    assert_equal(expect, table)
+  def test_add_args_to_url_arg_replaces_url_parameter
+    assert_equal("/abcdef?foo=bar&this=that",
+                 add_args_to_url("/abcdef?foo=wrong", foo: "bar", this: "that"))
   end
 
-  def test_make_table_row_without_columns
-    expect = "<table><tr>row without columns</tr></table>"
-    table = make_table(["row without columns"])
-    assert_equal(expect, table)
+  def test_add_args_to_url_append_args_to_url
+    assert_equal("/abcdef?a=2&foo=%22bar%22&this=that",
+                 add_args_to_url("/abcdef?foo=wrong&a=2",
+                                 foo: '"bar"', this: "that"))
+  end
+
+  def test_add_args_to_url_ending_with_id
+    assert_equal("/blah/blah/5?arg=new",
+                 add_args_to_url("/blah/blah/5", arg: "new"))
+  end
+
+  def test_add_args_to_url_id_arg_replaces_id_in_url
+    assert_equal("/blah/blah/4?arg=new",
+                 add_args_to_url("/blah/blah/5", arg: "new", id: 4))
+  end
+
+  def test_add_args_to_url_valid_utf_8_address_and_arg
+    assert_equal("/voilà?arg=a%C4%8D%E2%82%AC%CE%B5nt",
+                 add_args_to_url("/voilà", arg: "ač€εnt"))
+  end
+
+  def test_add_args_to_url_invalid_utf_8_address_and_arg
+    assert_equal("/blah\x80",
+                 add_args_to_url("/blah\x80", x: "foo\xA0"))
   end
 end
